backend/requirements.txt
fastapi
uvicorn
sqlalchemy
pydantic
azure-iot-device
requests
python-multipart
python-dotenv
